support Xcode5  64bit build
支持 xcode5 64位编译

+++ LKDBHelper/Helper/LKDB+Mapping.h
@@ -78,7 +78,7 @@ extern inline NSString* LKSQLTypeFromObjcType(NSString *objcType);
 
 -(id)initWithKeyMapping:(NSDictionary*)keyMapping propertyNames:(NSArray*)propertyNames propertyType:(NSArray*)propertyType primaryKeys:(NSArray*)primaryKeys;
 
-@property(readonly,nonatomic)int count;
+@property(readonly,nonatomic)NSUInteger count;
 @property(readonly,nonatomic)NSArray* primaryKeys;
 
 -(LKDBProperty*)objectWithIndex:(int)index;
